Import version v2.2

+++ include/sys/mman.h
@@ -1,12 +1,14 @@
 #ifndef _SYS_MMAN_H
+#include <stdint.h>
 #include <misc/sys/mman.h>
 
 #ifndef _ISOMAC
 /* Now define the internal interfaces.  */
+/* #1410 */
 extern void *__mmap (void *__addr, size_t __len, int __prot,
-		     int __flags, int __fd, __off_t __offset);
+		     uint64_t __flags, int __fd, __off_t __offset);
 extern void *__mmap64 (void *__addr, size_t __len, int __prot,
-		       int __flags, int __fd, __off64_t __offset);
+		       uint64_t __flags, int __fd, __off64_t __offset);
 extern int __munmap (void *__addr, size_t __len);
 extern int __mprotect (void *__addr, size_t __len, int __prot);
